eslint config and some code cleanup
