+++ .eslintrc.json
@@ -1,0 +1,7 @@
+{
+	"extends": "next",
+	"plugins": ["unused-imports"],
+	"rules": {
+		"unused-imports/no-unused-imports": "error"
+	}
+}
