+++ package.json
@@ -29,6 +29,7 @@
 	"devDependencies": {
 		"eslint": "^8.30.0",
 		"eslint-config-next": "13.0.3",
+		"eslint-plugin-unused-imports": "^2.0.0",
 		"prettier": "^2.8.4"
 	}
 }
