+++ src/components/Game/components/ConfirmModal.js
