+++ src/components/Game/components/index.js
@@ -1,0 +1,4 @@
+import ConfirmModal from './ConfirmModal'
+import StatusButton from './StatusButton'
+
+export { ConfirmModal, StatusButton }
