+++ src/components/components/index.js
@@ -1,7 +1,6 @@
 import AddGameModal from './AddGameModal'
 import AppBar from './AppBar'
 import NavTabs from './NavTabs'
-import ConfirmModal from './ConfirmModal'
 import GameCards from './GameCards'
 import LoadingOverlay from './Loading'
 import PlatformIcon from './PlatformIcon'
@@ -10,7 +9,6 @@ export {
 	AddGameModal,
 	AppBar,
 	NavTabs,
-	ConfirmModal,
 	GameCards,
 	LoadingOverlay,
 	PlatformIcon,
